Big-ass refactor
